temp

+++ src/app/board/post.service.ts
@@ -2,12 +2,15 @@ import { Injectable } from '@angular/core';
 import { Http, Response } from "@angular/http";
 import { Observable } from 'rxjs/Rx';
 import { JwtHttp } from 'ng2-ui-auth';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 import { environment } from '../../environments/environment';
 import { Post } from '../shared/models/post.model';
 import { PartialList } from '../shared/partial-list';
 
 import { HelperService } from '../core/helper.service';
+
+import { PostDeleteConfirmModalComponent } from './post-delete-confirm-modal/post-delete-confirm-modal.component';
 
 
 @Injectable()
@@ -17,7 +20,8 @@ export class PostService {
 
     constructor(
         private http: JwtHttp,
-        private helperService: HelperService
+        private helperService: HelperService,
+        private modalService: NgbModal
     ) { }
 
     get(id: string): Observable<Post> {
@@ -54,8 +58,14 @@ export class PostService {
     }
 
     remove(post: Post): Observable<boolean> {
-        return this.http.delete(this.postsUrl + post.id + '/').map((response: Response) => {
-            return true;
+        const modalRef = this.modalService.open(PostDeleteConfirmModalComponent);
+        return Observable.from(modalRef.result).flatMap(confirmed => {
+            if (confirmed) {
+                return this.http.delete(this.postsUrl + post.id + '/').map((response: Response) => {
+                    return true;
+                });
+            }
+            return Observable.of(false);
         });
     }
 }
